fix: LP64 ABI fixes — struct layouts, signal widening, lo/hi elimination
Fixes 31 sortix test regressions from the wasm64 LP64 migration:

- Widen si_value from i32 to i64 throughout signal path (RtSigEntry,
  dequeue, deliver, channel layout) for LP64 union sigval (8B sival_ptr)
- Fix siginfo_t construction in glue: use struct instead of hardcoded
  offsets (LP64 adds 4B padding after si_code before union)
- Fix sigtimedwait siginfo_t write offsets for LP64
- Fix SYS_RT_SIGQUEUEINFO to read si_value at LP64 offset 24 (8 bytes)
- Eliminate lo/hi splitting for LP64 syscalls: lseek, pread, pwrite,
  fallocate, ftruncate, truncate now receive full 64-bit values
- Fix ksigevent/sigevent struct reading for LP64 (union sigval at 8B
  shifts sigev_signo to offset 8, sigev_notify to offset 12)
- Fix F_GETLK/F_SETLK/F_SETLKW constants: LP64 uses 5/6/7 (native)
  instead of ILP32's 12/13/14 (64-bit lock variants)
- Fix SO_RCVTIMEO/SO_SNDTIMEO constants: LP64 uses 20/21 instead of
  ILP32's 66/67 (time64 variants)
- Add 12 missing wasm64posix musl arch files (ndbm, libintl_l, net_if,
  posix_getdents, posix_mem, siglongjmp, dcngettext)
- Rebuild dash with correct Linux signal name table (was using macOS
  BSD numbering: USR1=30 instead of 10)
- Remove ffsll from sortix XFAIL (passes on LP64 where long=64-bit)

Test results: cargo 703/0, vitest 252/0, libc-test 0 FAIL/22 XFAIL,
POSIX 0 FAIL/1 XFAIL, sortix 4815 PASS/0 FAIL/17 XFAIL

+++ crates/kernel/src/signal.rs
@@ -89,7 +89,8 @@ fn should_discard_pending(signum: u32, handler: &SignalHandler) -> bool {
 #[derive(Debug, Clone, Copy)]
 pub struct RtSigEntry {
     pub signum: u32,
-    pub si_value: i32,
+    /// Signal value (union sigval) — 8 bytes on LP64 to hold sival_ptr.
+    pub si_value: i64,
     /// SI_QUEUE (-1) if sent via sigqueue(), SI_USER (0) if via kill()/raise().
     pub si_code: i32,
 }
@@ -183,11 +184,11 @@ impl SignalState {
     }
 
     /// Mark a signal as pending with an si_value (for sigqueue — SI_QUEUE).
-    pub fn raise_with_value(&mut self, signum: u32, si_value: i32) -> bool {
+    pub fn raise_with_value(&mut self, signum: u32, si_value: i64) -> bool {
         self.raise_internal(signum, si_value, -1) // SI_QUEUE
     }
 
-    fn raise_internal(&mut self, signum: u32, si_value: i32, si_code: i32) -> bool {
+    fn raise_internal(&mut self, signum: u32, si_value: i64, si_code: i32) -> bool {
         if signum == 0 || signum >= 65 {
             return false;
         }
@@ -243,9 +244,9 @@ impl SignalState {
     /// For RT signals, removes one queued instance; clears pending bit only when
     /// no more instances remain. For standard signals, clears the pending bit.
     /// Returns (si_value, si_code) of the consumed signal instance.
-    pub fn consume_one(&mut self, signum: u32) -> (i32, i32) {
+    pub fn consume_one(&mut self, signum: u32) -> (i64, i32) {
         if signum == 0 || signum >= NSIG { return (0, 0); }
-        let (mut si_value, mut si_code) = (0i32, 0i32);
+        let (mut si_value, mut si_code) = (0i64, 0i32);
         if let Some(pos) = self.rt_queue.iter().position(|e| e.signum == signum) {
             si_value = self.rt_queue[pos].si_value;
             si_code = self.rt_queue[pos].si_code;
@@ -289,14 +290,14 @@ impl SignalState {
     /// RT signals (32-63) are dequeued from the queue; the pending bit is
     /// only cleared when no more instances of that signal remain in the queue.
     /// Returns (signum, si_value, si_code). si_value and si_code are 0 for standard signals.
-    pub fn dequeue(&mut self) -> Option<(u32, i32, i32)> {
+    pub fn dequeue(&mut self) -> Option<(u32, i64, i32)> {
         let deliverable = self.pending & !self.blocked;
         if deliverable == 0 {
             return None;
         }
         // trailing_zeros gives 0-based bit position; signal number = bit + 1
         let signum = deliverable.trailing_zeros() + 1;
-        let (mut si_value, mut si_code) = (0i32, 0i32);
+        let (mut si_value, mut si_code) = (0i64, 0i32);
         if signum >= SIGRTMIN {
             // RT signal: dequeue one instance from the queue
             if let Some(pos) = self.rt_queue.iter().position(|e| e.signum == signum) {
